More code cleanup. Pulled GUI objects into a new GraphicalUserInterface class. Added logging capability.

+++ Empire/Model/Planet.cs
@@ -7,11 +7,13 @@ using System.Threading.Tasks;
 
 namespace Empire.Model
 {
+    // A very simple class to distinguish planets from other game entities.
+    // Maybe someday planets will have populations, sovereignty, exports, etc.
     class Planet : Entity
     {
         public Planets planetID { get; private set; }
 
-        public Planet(int x = 0, int y = 0, Planets ID = Planets.planet1) : base(x, y) {
+        public Planet(float x = 0, float y = 0, Planets ID = Planets.planet1) : base(x, y) {
             Location = new Vector2(x, y);
             planetID = ID;
             this.Type = EntityType.Planet;
